Updating and splitting Functions Chapter.
